Time: 129 ms (67.44%) | Memory: 122.5 MB (44.57%) - LeetSync

+++ 3213-count-subarrays-where-max-element-appears-at-least-k-times/count-subarrays-where-max-element-appears-at-least-k-times.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long countSubarrays(vector<int>& nums, int k) {
         int mx = 0, sz = nums.size();
-        long ans = 0;
+        long long ans = 0;
         for(int i = 0;i<sz;++i)
             mx = max(mx, nums[i]);
         int left = 0, right = 0, cnt = 0;
@@ -11,8 +11,8 @@ public:
             if(nums[right] == mx)
                 el.push_back(right), ++cnt;
             if(cnt == k){
-                int before = el.front() - left, after = sz - right - 1;
-                ans += before + after + (long)before * after + 1;
+                long long before = el.front() - left, after = sz - right - 1;
+                ans += before + after + before * after + 1;
                 while(nums[left++] != mx);
                 --cnt;
                 el.pop_front();
